+++ sdk/servicebus/azure-servicebus/tests/test_sessions.py
@@ -11,6 +11,7 @@ import os
 import pytest
 import time
 import uuid
+import pickle
 from datetime import timedelta
 
 from azure.servicebus import (
@@ -700,6 +701,8 @@ class TestServiceBusSession(AzureMgmtRecordedTestCase):
                 results.append(renewable)
 
             renewer = AutoLockRenewer()
+            if not uamqp_transport:
+                renewer = pickle.loads(pickle.dumps(renewer))
             messages = []
             with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=10, receive_mode=ServiceBusReceiveMode.PEEK_LOCK, prefetch_count=10) as receiver:
                 renewer.register(receiver,